+++ server.js
@@ -8,13 +8,19 @@
 'use strict';
 const express = require('express');
 const router = require('./app/router.js');
-const bodyParser = require('body-parser');
 const config = require('./config.json');
 const app = express();
 const port = process.env.PORT || config.port || 80;
-
 console.log('Listening to port: ' + port);
-
-app.use(bodyParser.json());
+app.use((req, res, next) => {
+	req.rawBody = '';
+	req.on('data', chunk => {
+		req.rawBody += chunk;
+	});
+	req.on('end', () => {
+		req.body = JSON.parse(req.rawBody);
+		next();
+	});
+})
 app.use('/', router); // Register router
 app.listen(port); // Start server